repo: drop test-only git2 merge wrapper
